+++ src/main/java/br/pucrio/les/esg_token_backend/controllers/ValueController.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
@@ -25,5 +27,11 @@ public class ValueController {
     @ResponseStatus(code = HttpStatus.FOUND)
     List<Value> index() {
         return this.valueService.index();
+    }
+
+    @PostMapping
+    @ResponseStatus(code = HttpStatus.CREATED)
+    Value create(@RequestBody Value value) {
+        return this.valueService.create(value);
     }
 }